feat(video): Implement automatic audio track selection and fix track removal detection
This change introduces a new feature for automatic audio track selection based on language priority.

- The priority is configurable via the `PREFERRED_LANGUAGES` environment variable, with a default of Telugu -> Hindi -> English.
- Language detection is enhanced to support 2-letter and 3-letter codes, as well as native script names in stream metadata.
- If no preferred language is found, all audio tracks are kept.

This change also fixes a bug where manual track removal was not being detected correctly.
- The results of manual stream selection are now properly recorded.
- The automatic processing logic is prevented from overriding manual selections.
- Status reporting for video processing has been made more robust.

This also fixes an `ImportError` for `get_metavideo` caused by a circular dependency.

v2.2-debug
