+++ bot/helper/ext_utils/media_utils.py
@@ -496,3 +496,24 @@ class GenSS:
     async def ddl_ss(self):
         self._name = get_url_name(self._path)
         await self.file_ss()
+
+
+async def get_metavideo(url):
+    """Get media metadata from a URL using ffprobe."""
+    try:
+        process = await create_subprocess_exec(
+            'ffprobe', '-hide_banner', '-loglevel', 'error', '-print_format', 'json',
+            '-show_format', url,
+            stdout=PIPE, stderr=PIPE
+        )
+        stdout, stderr = await process.communicate()
+        if process.returncode != 0:
+            LOGGER.error(f"ffprobe error for URL {url}: {stderr.decode().strip()}")
+            return None, None
+        media_info = literal_eval(stdout.decode().strip())
+        duration = media_info.get('format', {}).get('duration', 0)
+        size = media_info.get('format', {}).get('size', 0)
+        return duration, {'size': size}
+    except Exception as e:
+        LOGGER.error(f"Exception while getting media metadata for URL {url}: {e}")
+        return None, None
