+++ log.txt
@@ -4,3 +4,6 @@
 14-Aug-25 07:38:50 AM: [WARNING: __init__.py - 316] ~ MEGA Credentials not provided!
 14-Aug-25 07:38:50 AM: [WARNING: __init__.py - 760] ~ PORT is not set or not a valid integer. Skipping gunicorn start.
 14-Aug-25 07:38:52 AM: [INFO: __init__.py - 807] ~ Creating client Pyrofork V2.2.11...
+14-Aug-25 07:50:02 AM: [WARNING: __init__.py - 316] ~ MEGA Credentials not provided!
+14-Aug-25 07:50:02 AM: [WARNING: __init__.py - 760] ~ PORT is not set or not a valid integer. Skipping gunicorn start.
+14-Aug-25 07:50:04 AM: [INFO: __init__.py - 807] ~ Creating client Pyrofork V2.2.11...
