+++ qBittorrent/data/logs/qbittorrent.log
@@ -50,3 +50,29 @@
 (C) 2025-08-14T07:38:51 - WebUI: Unable to bind to IP: *, port: 8090. Reason: The bound address is already in use
 (N) 2025-08-14T07:38:51 - qBittorrent termination initiated
 (N) 2025-08-14T07:38:56 - qBittorrent is now ready to exit
+(N) 2025-08-14T07:50:02 - qBittorrent v4.6.3 started
+(N) 2025-08-14T07:50:02 - Using config directory: /app/qBittorrent/config
+(N) 2025-08-14T07:50:02 - qBittorrent termination initiated
+(N) 2025-08-14T07:50:02 - qBittorrent is now ready to exit
+(N) 2025-08-14T07:50:02 - qBittorrent v4.6.3 started
+(N) 2025-08-14T07:50:02 - Using config directory: /app/qBittorrent/config
+(N) 2025-08-14T07:50:02 - Trying to listen on the following list of IP addresses: "0.0.0.0:47671,[::]:47671"
+(I) 2025-08-14T07:50:02 - Peer ID: "-qB4630-"
+(I) 2025-08-14T07:50:02 - HTTP User-Agent: "qBittorrent/4.6.3"
+(I) 2025-08-14T07:50:02 - Distributed Hash Table (DHT) support: ON
+(I) 2025-08-14T07:50:02 - Local Peer Discovery support: ON
+(I) 2025-08-14T07:50:02 - Peer Exchange (PeX) support: ON
+(I) 2025-08-14T07:50:02 - Anonymous mode: OFF
+(I) 2025-08-14T07:50:02 - Encryption support: ON
+(W) 2025-08-14T07:50:02 - Embedded Tracker: Unable to bind to IP: 0.0.0.0, port: 9000. Reason: The bound address is already in use
+(I) 2025-08-14T07:50:02 - UPnP/NAT-PMP support: ON
+(I) 2025-08-14T07:50:02 - Successfully listening on IP. IP: "127.0.0.1". Port: "TCP/47671"
+(I) 2025-08-14T07:50:02 - Successfully listening on IP. IP: "127.0.0.1". Port: "UTP/47671"
+(I) 2025-08-14T07:50:02 - Successfully listening on IP. IP: "192.168.0.2". Port: "TCP/47671"
+(I) 2025-08-14T07:50:02 - Successfully listening on IP. IP: "192.168.0.2". Port: "UTP/47671"
+(W) 2025-08-14T07:50:02 - Couldn't load IP geolocation database. Reason: No such file or directory
+(N) 2025-08-14T07:50:03 - Using built-in WebUI.
+(W) 2025-08-14T07:50:03 - Couldn't load WebUI translation for selected locale (C).
+(C) 2025-08-14T07:50:03 - WebUI: Unable to bind to IP: *, port: 8090. Reason: The bound address is already in use
+(N) 2025-08-14T07:50:03 - qBittorrent termination initiated
+(N) 2025-08-14T07:50:08 - qBittorrent is now ready to exit
